Modify Section pre account refactor

+++ FinTrac/BusinessLogic/Account Components/CreditCardAccount.cs
@@ -21,7 +21,7 @@ namespace BusinessLogic.Account_Components
         #region Constructor
         public CreditCardAccount() { }
 
-        public CreditCardAccount(string name, CurrencyEnum currency, string issuingBank, string last4Digits, int availableCredit, DateTime closingDate) : base(name, currency)
+        public CreditCardAccount(string name, CurrencyEnum currency, string issuingBank, string last4Digits, decimal availableCredit, DateTime closingDate) : base(name, currency)
         {
             IssuingBank = issuingBank;
             Last4Digits = last4Digits;
